Finished server pool; added test

+++ src/TetrinetServerPool.cpp
@@ -2,12 +2,14 @@
 
 #include <boost/format.hpp>
 #include <boost/lexical_cast.hpp>
+#include <boost/filesystem.hpp>
 #include <algorithm>
 #include <signal.h>
 #include <stdexcept>
 
 using namespace std;
 using namespace boost;
+using namespace boost::filesystem;
 
 TetrinetServerPool::TetrinetServerPool(const std::string poolName, const unsigned poolSize, const uint32_t ipv4, const uint16_t portStart, const std::string pidFileExpr, const std::string startCmdExpr, const std::string stopCmdExpr)
   : poolName(poolName),
@@ -59,6 +61,12 @@ void TetrinetServerPool::RunMatch(Match& match)
   (*find_if(servers.begin(), servers.end(), [](const TetrinetServer* i) { return i->GetState() == TetrinetServerState::VACANT; }))->SetMatch(match);
 }
 
+void TetrinetServerPool::StartServers()
+{
+  for (TetrinetServer* server : servers)
+    server->Start();
+}
+
 TetrinetServerPool::TetrinetServer::TetrinetServer(const uint16_t port, const std::string pidFile, const std::string startCmd, const std::string stopCmd)
   : port(port),
     pid(-1),
@@ -76,6 +84,8 @@ TetrinetServerPool::TetrinetServer::~TetrinetServer()
 {
   if (!pidStream.is_open())
     pidStream.close();
+  if (GetState() != TetrinetServerState::STOPPED)
+    Stop();
 }
 
 const uint16_t TetrinetServerPool::TetrinetServer::GetPort() const
@@ -85,10 +95,12 @@ const uint16_t TetrinetServerPool::TetrinetServer::GetPort() const
 
 const TetrinetServerState TetrinetServerPool::TetrinetServer::GetState() const
 {
+  if (!exists(pidFile))
+    return TetrinetServerState::STOPPED;
   return state = (kill(pid, 0) == 0 ? (match != none ? TetrinetServerState::VACANT : TetrinetServerState::OCCUPIED) : TetrinetServerState::STOPPED);
 }
 
-const optional<Match&> TetrinetServerPool::TetrinetServer::GetMatch() const
+const optional<const Match&> TetrinetServerPool::TetrinetServer::GetMatch() const
 {
   return match;
 }
@@ -100,13 +112,15 @@ void TetrinetServerPool::TetrinetServer::SetMatch(const Match& match)
   
   this->match = match;
   state = TetrinetServerState::OCCUPIED;
-  onCompleteConnection = match.addOnComplete(Match::CompleteSignal::slot_type(bind(&TetrinetServerPool::TetrinetServer::ClearMatch, this, _1)));
+  onCompleteConnection = match.AddOnComplete(Match::CompleteSignal::slot_type(bind(&TetrinetServerPool::TetrinetServer::ClearMatch, this, _1)));
 }
 
 void TetrinetServerPool::TetrinetServer::ClearMatch(const Match& match)
 {
-  if (!this->match && optional<const Match&>(const_cast<const Match&>(this->match.get())) != optional<const Match&>(match)) // this whole line is derp
-    throw runtime_error("Wrong match event received, or no match in progress");
+  if (!this->match)
+    throw runtime_error("No match in progress");
+  if (&(this->match.value()) != &match)
+    throw runtime_error("Wrong match event received");
   
   state = TetrinetServerState::VACANT;
   this->match = none;
@@ -121,7 +135,9 @@ void TetrinetServerPool::TetrinetServer::Start()
 {
   if (!pidStream.is_open())
     pidStream.close();
-  system(startCmd.c_str());
+  std::system(startCmd.c_str());
+  if (!exists(pidFile))
+    throw runtime_error("PID file does not exist");
   pidStream.open(pidFile);
 
   string pidS;
@@ -131,5 +147,5 @@ void TetrinetServerPool::TetrinetServer::Start()
 
 void TetrinetServerPool::TetrinetServer::Stop()
 {
-  system(stopCmd.c_str());
+  std::system(stopCmd.c_str());
 }
